requirements.txt
openai==0.27.8
python-dotenv==0.20.0
langchain==0.0.205